fix: make footer responsive
edited footer.js
edited footer styles in App.js

fixed #13
