+++ src/app.js
@@ -4,18 +4,18 @@ import NavBar from './navbar/NavBar';
 import './App.css';
 import LandingPage from './pages/LandingPage';
 import Provider from './context/Context';
-import Footer from './intro/Footer';
+import Footer from './footer/Footer';
 
 class App extends Component {
-  render() {
-    return (
-      <Provider>
-        <NavBar />
-        <LandingPage />
-        <Footer />
-      </Provider>
-    );
-  }
+	render() {
+		return (
+			<Provider>
+				<NavBar />
+				<LandingPage />
+				<Footer />
+			</Provider>
+		);
+	}
 }
 
 export default App;
